Store bool values as int32 values.
This kind of sucks. The problem is that the V1 vdf format overloads
int32 fields for storing boolean values. As a result, the parser
only knows that it is reading in an int32 field. This means that
the library must also store boolean values as int32. Using two
separate fields would be better, but that would require implementing
some means to identify which value to use. E.g.,

if version == V1 {
	return int32ToBool(o.int32Value)
} else {
	return o.boolValue
}

I am leaving this alone for now. API changes might be required to
support such functionality.